fix(ci): grant permissions to close master PRs
The `close-master-pr` workflow was failing with a "Resource not
accessible by integration" error. This was caused by the default
`GITHUB_TOKEN` not having sufficient permissions to close pull requests
when using the `pull_request_target` trigger.

This change adds the `pull-requests: write` permission to the job,
allowing the `superbrothers/close-pull-request` action to function
correctly.